+++ frontend/src/components/UpdateInterventionForm/UpdateInterventionForm.js
@@ -224,7 +224,7 @@ class UpdateInterventionForm extends Component {
                             Annuler
                         </button>
                         <button type="submit" className={styles.submitButton}>
-                            <i className="fa-solid fa-save"></i>
+                            <i className="fa-regular fa-save"></i>
                             Enregistrer
                         </button>
                     </div>