+++ frontend/src/pages/Statistics/StatisticsPage.js
@@ -1,0 +1,30 @@
+import React, { Component } from "react";
+import { useNavigate } from "react-router-dom";
+import TemplateGlobal from "../Template/TemplateGlobal";
+import styles from "./StatisticsPage.module.css";
+
+// Composant fonctionnel wrapper
+const CompanyPageWrapper = () => {
+    const navigate = useNavigate();
+    return <StatisticsPage navigate={navigate} location={location} />;
+};
+
+class StatisticsPage extends Component {
+    constructor(props) {
+        super(props);
+        this.state = {};
+    }
+
+    render() {
+        return (
+            <>
+                <TemplateGlobal />
+                <div className={styles.container}>
+                    <h1 className={styles.pageTitle}>Statistiques</h1>
+                </div>
+            </>
+        );
+    }
+}
+
+export default CompanyPageWrapper;