+++ CHANGELOG.md
@@ -10,10 +10,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 -   Add company name
 -   Add basic notification for create events
--   Add statistics page
+-   Add statistics page with basic graphs
 -   Add current month event stats card
 -   Add events by employee stats card
 -   Add download document link
+-   Add tel and mail link to table clients
 
 ## Fixes